+++ catalog/app/website/components/Backgrounds/Overlay2.js
@@ -1,0 +1,17 @@
+import { unstable_Box as Box } from '@material-ui/core/Box'
+import { styled } from '@material-ui/styles'
+
+import offset from 'utils/bgOffset'
+
+import overlay2 from './overlay2.png'
+
+export default styled(Box)({
+  backgroundImage: `url(${overlay2})`,
+  backgroundPosition: `top left ${offset(0)}`,
+  backgroundSize: 'cover',
+  height: 1105,
+  left: 0,
+  mixBlendMode: 'overlay',
+  position: 'absolute',
+  right: 0,
+})
