--- catalog/app/website/components/Backgrounds/index.js
@@ -1,1 +1,0 @@
-export * from './Backgrounds'
