+++ catalog/app/website/components/Layout.js
@@ -4,10 +4,8 @@ import { ThemeProvider } from '@material-ui/styles'
 import * as Layout from 'components/Layout'
 import createTheme from 'website/theme'
 
-export { default as Container } from 'components/Layout/Container'
-
-export default () => (
+export default ({ children, ...props }) => (
   <ThemeProvider theme={createTheme}>
-    <Layout.Layout pre={children} />
+    <Layout.Layout pre={children} {...props} />
   </ThemeProvider>
 )
