+++ catalog/app/website/pages/Landing.js
@@ -1,9 +1,7 @@
 import * as React from 'react'
 
+import Hero from 'website/components/Hero'
 import Layout from 'website/components/Layout'
-
-import Theme from 'website/theme'
-import Hero from './Hero'
 //import Highlights from './Highlights'
 //import MoreAboutQuilt from './MoreAboutQuilt'
 //import Pricing from './Pricing'
